# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GuardBench API
  version: 1.0.0
  description: |
    Status: APPROVED
    Owner: Backend
    Last reviewed: 2026-09-25
    Canonical source: GitHub
    Origin: https://app.notion.com/p/3c0eeed6b62d805dac0be8db487b1359
    HTTP 상태와 GuardBench의 실행·판정 상태는 서로 독립적이다.
    이 schema는 ADR 0013의 합의된 AI Application Target + Response Behavior Classifier 계약을 명세한다.
    #115/#125/#128의 OpenAI-compatible HTTP Application Target 계약, #173의 classifier 전환,
    #117의 Worker orchestration과 결과 경계는 구현되었고,
    #118의 현재 Run Quality Gate가 구현되었으며 #119가 Regression을 담당한다.
servers:
  - url: http://localhost:8080
security: []
...
paths:
...
  /api/v1/test-runs:
    post:
      tags: [TestRun]
      operationId: createTestRun
      summary: TestRun 비동기 실행 요청
      description: |
        실행 완료가 아니라 TestRun 실행 요청을 안전하게 접수한다.
        요청은 하나의 OpenAI-compatible HTTP AI Application Target만 포함한다.
        Classifier provider, endpoint, model과 prompt는 서버 deployment configuration이 관리한다.
        실행 조건과 현재 TestCase의 불변 Snapshot을 고정하며 HTTP Target의 identifier와 model도 실행 조건으로 고정한다.
        커밋 후 발생한 SQS 발행·Application 실행·Evaluator 오류는 이 HTTP 요청의 실패로 바꾸지 않고 TestRun 결과로 기록한다.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestRunCreateReq' }
            example:
              testSuiteId: 1
              target:
                type: HTTP_ENDPOINT
                identifier: https://example.com/v1/chat/completions
                revision: v2
                model: example-model
              qualityGatePolicy:
                assertionPassRateThreshold: 0.9
                executionSuccessRateThreshold: 0.98
      responses:
        '202':
          description: |
            실행 요청이 접수되었다.
            같은 Idempotency-Key와 같은 요청을 재전송한 경우 새 TestRun을 만들지 않고 기존 TestRun과 현재 상태를 반환한다.
          headers:
            Location:
              description: 생성되었거나 기존에 접수된 TestRun의 상세 조회 경로
              schema:
                type: string
                example: /api/v1/test-runs/901
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestRunCreateApiResponse' }
              example:
                httpStatus: 202
                message: TestRun 실행 요청이 접수되었습니다.
                data:
                  id: 901
                  testSuiteId: 1
                  status: QUEUED
                  testCaseCount: 253
                  target:
                    type: HTTP_ENDPOINT
                    identifier: https://example.com/v1/chat/completions
                    revision: v2
                    model: example-model
                  createdAt: '2026-08-24T14:30:00Z'
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/TestSuiteNotFound' }
        '409': { $ref: '#/components/responses/TestRunCreateConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [TestRun]
      operationId: listTestRuns
      summary: TestRun 목록 조회
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/TestRunSort'
        - $ref: '#/components/parameters/TestRunTestSuiteId'
        - $ref: '#/components/parameters/TestRunStatusFilter'
        - $ref: '#/components/parameters/TestRunExecutionOutcomeFilter'
        - $ref: '#/components/parameters/TestRunQualityGateStatusFilter'
        - $ref: '#/components/parameters/TestRunCreatedFrom'
        - $ref: '#/components/parameters/TestRunCreatedTo'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestRunListApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        TestRun 중복 생성을 방지하는 선택 키다. UUID 사용을 권장한다.
        같은 키와 같은 요청은 기존 TestRun을 반환하고, 같은 키를 다른 요청에 재사용하면 409로 응답한다.
        생략하면 요청할 때마다 새로운 TestRun을 생성한다.
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: 31c83d18-12c4-47b7-9ed4-23e621cb9999
...
    PageNumber:
      name: page
      in: query
      description: |
        1부터 시작하는 Offset 페이지 번호다.
        실제 마지막 페이지를 초과하면 200 OK와 빈 items를 반환하며 요청 번호를 응답에 유지한다.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: size
      in: query
      description: 페이지당 항목 수. 1~100 범위이며 범위를 벗어나면 400으로 응답한다.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    TestRunSort:
      name: sort
      in: query
      description: |
        TestRun 다중 정렬 조건. 허용 필드는 createdAt, startedAt, completedAt, updatedAt, testCaseCount, id다.
        생략하면 createdAt,desc와 id,desc를 적용하고 요청에 id가 없으면 마지막에 id,desc를 추가한다.
        startedAt과 completedAt의 null은 정렬 방향과 무관하게 항상 마지막이다.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          pattern: '^(createdAt|startedAt|completedAt|updatedAt|testCaseCount|id),(asc|desc)$'
      example: ['createdAt,desc', 'id,desc']
    TestRunTestSuiteId:
      name: testSuiteId
      in: query
      description: TestSuite ID 정확히 일치 Filter
      schema: { type: integer, format: int64, minimum: 1 }
    TestRunStatusFilter:
      name: status
      in: query
      description: 같은 Parameter를 반복하면 값들을 OR로 결합한다.
      style: form
      explode: true
      schema:
        type: array
        items: { $ref: '#/components/schemas/TestRunStatus' }
    TestRunExecutionOutcomeFilter:
      name: executionOutcome
      in: query
      description: |
        같은 Parameter를 반복하면 값들을 OR로 결합한다.
        executionOutcome이 null인 TestRun은 이 Filter와 일치하지 않는다.
      style: form
      explode: true
      schema:
        type: array
        items: { $ref: '#/components/schemas/TestRunExecutionOutcome' }
    TestRunQualityGateStatusFilter:
      name: qualityGateStatus
      in: query
      description: |
        같은 Parameter를 반복하면 값들을 OR로 결합한다.
        Quality Gate가 아직 null인 TestRun은 이 Filter와 일치하지 않는다.
      style: form
      explode: true
      schema:
        type: array
        items: { $ref: '#/components/schemas/QualityGateStatus' }
    TestRunCreatedFrom:
      name: createdFrom
      in: query
      description: 생성 시각 하한이며 해당 시각을 포함한다. 생략하면 과거 방향에 제한을 두지 않는다.
      schema: { type: string, format: date-time }
    TestRunCreatedTo:
      name: createdTo
      in: query
      description: 생성 시각 상한이며 해당 시각은 포함하지 않는다. 생략하면 미래 방향에 제한을 두지 않는다.
      schema: { type: string, format: date-time }
...
  responses:
    ValidationError:
      description: 요청값 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorApiResponse' }
          example:
            httpStatus: 400
            message: 요청 값이 올바르지 않습니다.
            data:
              code: VALIDATION_ERROR
              errors:
                - field: name
                  message: 이름은 필수입니다.
    TestSuiteNotFound:
      description: TestSuite를 찾을 수 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorApiResponse' }
          example:
            httpStatus: 404
            message: TestSuite를 찾을 수 없습니다.
            data: { code: TEST_SUITE_NOT_FOUND }
...
    TestRunCreateConflict:
      description: TestRun을 접수할 수 없는 현재 상태 또는 Idempotency-Key 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorApiResponse' }
          examples:
            emptyTestSuite:
              summary: 실행 가능한 TestCase가 없음
              value:
                httpStatus: 409
                message: 실행 가능한 TestCase가 없습니다.
                data: { code: TEST_SUITE_EMPTY }
            idempotencyKeyConflict:
              summary: 같은 키를 다른 요청에 재사용
              value:
                httpStatus: 409
                message: Idempotency-Key가 다른 요청에 이미 사용되었습니다.
                data: { code: IDEMPOTENCY_KEY_CONFLICT }
...
    InternalServerError:
      description: 요청 접수 트랜잭션을 완료하지 못한 예상하지 못한 서버 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorApiResponse' }
          example:
            httpStatus: 500
            message: 서버 내부 오류가 발생했습니다.
            data: { code: INTERNAL_SERVER_ERROR }
  schemas:
...
    PageMetaRes:
      type: object
      description: |
        Offset Pagination 메타데이터다. 범위를 초과한 페이지도 number에는 요청 번호를 유지한다.
        필터 결과가 0건이면 totalElements와 totalPages는 모두 0이다.
      required: [number, size, totalElements, totalPages, hasPrevious, hasNext]
      properties:
        number: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasPrevious: { type: boolean }
        hasNext: { type: boolean }
...
    TargetReferenceReq:
      type: object
      description: |
        TestRun이 실행할 단일 AI Application Target이다.
        MVP는 OpenAI-compatible chat completions HTTP/HTTPS endpoint만 허용한다.
      required: [type, identifier, model]
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [HTTP_ENDPOINT]
        identifier:
          type: string
          format: uri
          minLength: 1
          pattern: '^https?://.+'
          description: OpenAI-compatible chat completions AI Application의 full HTTP 또는 HTTPS 호출 URL
        revision:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          description: 배포 버전, 모델 버전 또는 commit처럼 사용자가 Application revision을 식별하는 선택 문자열
        model:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          description: OpenAI-compatible chat completions request body에 전달할 필수 모델 식별자
    TestRunCreateReq:
      type: object
      description: |
        TestSuite의 현재 TestCase 전체를 실행 대상으로 사용하며 TestCase ID 목록은 받지 않는다.
        TestCase가 없으면 409 TEST_SUITE_EMPTY로 응답한다.
        classifier 설정은 서버 deployment configuration으로 관리하며 요청에서 받지 않는다.
        qualityGatePolicy를 생략하면 두 threshold 모두 0.95를 적용한다.
      required: [testSuiteId, target]
      additionalProperties: false
      properties:
        testSuiteId: { type: integer, format: int64, minimum: 1 }
        target: { $ref: '#/components/schemas/TargetReferenceReq' }
        qualityGatePolicy:
          $ref: '#/components/schemas/QualityGatePolicyReq'
    QualityGatePolicyReq:
      type: object
      required: [assertionPassRateThreshold, executionSuccessRateThreshold]
      additionalProperties: false
      description: TestRun 생성 시 고정되어 해당 Run의 최종 판정에 사용되는 Quality Gate 정책
      properties:
        assertionPassRateThreshold:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 평가 가능한 Assertion 중 PASS 비율의 최소 통과 기준
        executionSuccessRateThreshold:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 전체 Snapshot 중 성공한 실행 비율의 최소 통과 기준
    TestRunStatus:
      type: string
      description: |
        QUEUED는 접수 트랜잭션 완료, PREPARING은 Target 준비와 대상 고정,
        RUNNING은 테스트 실행 중, FINISHED는 정상 완료 또는 오류로 처리가 종료된 상태다.
      enum: [QUEUED, PREPARING, RUNNING, FINISHED]
    TestRunCreateRes:
      type: object
      description: |
        새 접수는 QUEUED를 반환한다. 같은 Idempotency-Key와 같은 요청의 재전송은 기존 TestRun의 현재 status를 반환할 수 있다.
        실행 전에는 의미가 없는 executionOutcome과 qualityGateStatus는 포함하지 않는다.
      required: [id, testSuiteId, status, testCaseCount, target, createdAt]
      additionalProperties: false
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        testSuiteId: { type: integer, format: int64, minimum: 1 }
        status: { $ref: '#/components/schemas/TestRunStatus' }
        testCaseCount:
          type: integer
          minimum: 1
          description: 접수 트랜잭션에서 고정된 TestCaseSnapshot 개수
        target: { $ref: '#/components/schemas/TargetReferenceRes' }
        createdAt: { type: string, format: date-time, description: UTC 기준 ISO 8601 접수 시각 }
    TestRunExecutionOutcome:
      type: string
      enum: [COMPLETED, ERROR, INCOMPLETE]
    NullableTestRunExecutionOutcome:
      type: string
      nullable: true
      description: TestRunExecutionOutcome과 같은 값 집합에 아직 결정되지 않았음을 뜻하는 null을 더한 표현이다.
      enum: [COMPLETED, ERROR, INCOMPLETE, null]
    QualityGateStatus:
      type: string
      enum: [PASS, FAIL, NOT_EVALUATED]
    NullableQualityGateStatus:
      type: string
      nullable: true
      description: QualityGateStatus와 같은 값 집합에 아직 결정되지 않았음을 뜻하는 null을 더한 표현이다.
      enum: [PASS, FAIL, NOT_EVALUATED, null]
    TestRunListItemRes:
      type: object
      description: |
        실행 이력 목록용 요약이다. Application Target revision은 포함하지 않는다.
        executionOutcome, qualityGateStatus와 qualityGateMetrics는 아직 결정되지 않았으면 null이다.
        Quality Gate가 NOT_EVALUATED이면 qualityGateMetrics도 null이다.
      required:
        - id
        - testSuiteId
        - status
        - testCaseCount
        - progress
        - executionOutcome
        - qualityGateStatus
        - qualityGateMetrics
        - createdAt
        - startedAt
        - completedAt
        - updatedAt
      additionalProperties: false
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        testSuiteId: { type: integer, format: int64, minimum: 1 }
        status: { $ref: '#/components/schemas/TestRunStatus' }
        testCaseCount: { type: integer, minimum: 1 }
        progress: { $ref: '#/components/schemas/TestRunProgressRes' }
        executionOutcome: { $ref: '#/components/schemas/NullableTestRunExecutionOutcome' }
        qualityGateStatus: { $ref: '#/components/schemas/NullableQualityGateStatus' }
        qualityGateMetrics: { $ref: '#/components/schemas/QualityGateMetricsRes' }
        createdAt: { type: string, format: date-time, description: UTC 기준 ISO 8601 접수 시각 }
        startedAt:
          type: string
          format: date-time
          nullable: true
          description: PREPARING 시작 시각. null 정렬은 방향과 무관하게 마지막이다.
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: FINISHED 도달 시각. null 정렬은 방향과 무관하게 마지막이다.
        updatedAt: { type: string, format: date-time, description: 상태 또는 진행률의 마지막 변경 시각 }
    TestRunListRes:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TestRunListItemRes' }
        page: { $ref: '#/components/schemas/PageMetaRes' }
    TestRunProgressRes:
      type: object
      description: |
        Snapshot의 단일 Target 처리가 터미널 상태가 되면 처리 완료 한 건으로 계산한다.
        실패와 timeout도 더 이상 재시도하지 않는 터미널 상태라면 processedTestCaseCount에 포함한다.
      required: [processedTestCaseCount, percent]
      additionalProperties: false
      properties:
        processedTestCaseCount: { type: integer, minimum: 0 }
        percent: { type: number, format: double, minimum: 0, maximum: 100 }
    TargetReferenceRes:
      type: object
      required: [type, identifier, revision, model]
      additionalProperties: false
      properties:
        type: { type: string, enum: [HTTP_ENDPOINT] }
        identifier:
          type: string
          format: uri
          pattern: '^https?://.+'
        revision:
          type: string
          nullable: true
          description: 요청한 Application revision. 요청에서 생략하면 null이다.
        model:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          description: OpenAI-compatible 요청에 사용한 필수 모델 식별자
...
    QualityGateMetricsRes:
      type: object
      nullable: true
      required: [assertionPassRate, executionSuccessRate, assertion, execution]
      additionalProperties: false
      description: 현재 TestRun의 Assertion 및 실행 집계 지표. 미결정 또는 NOT_EVALUATED이면 null이다.
      properties:
        assertionPassRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          deprecated: true
          description: 기존 소비처 호환용 Assertion PASS 비율. assertion.value와 같은 값이다.
        executionSuccessRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          deprecated: true
          description: 기존 소비처 호환용 실행 성공 비율. execution.value와 같은 값이다.
        assertion:
          $ref: '#/components/schemas/QualityGateMetricRes'
        execution:
          $ref: '#/components/schemas/QualityGateMetricRes'
    QualityGateMetricRes:
      type: object
      required: [value, threshold, passed]
      additionalProperties: false
      description: Quality Gate 판정 당시 사용한 단일 지표의 값과 기준, 통과 여부
      properties:
        value:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 판정에 사용한 현재 비율
        threshold:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 판정 당시 적용한 최소 통과 기준
        passed:
          type: boolean
          description: backend가 판정한 이 지표의 통과 여부
...
    ErrorDetail:
      type: object
      required: [code]
      properties:
        code: { type: string, example: TEST_RUN_NOT_FOUND }
    FieldErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ValidationErrorDetail:
      type: object
      required: [code, errors]
      properties:
        code: { type: string, enum: [VALIDATION_ERROR] }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldErrorDetail' }
    ErrorApiResponse:
      type: object
      required: [httpStatus, message, data]
      properties:
        httpStatus: { type: integer }
        message: { type: string }
        data: { $ref: '#/components/schemas/ErrorDetail' }
    ValidationErrorApiResponse:
      type: object
      required: [httpStatus, message, data]
      properties:
        httpStatus: { type: integer, enum: [400] }
        message: { type: string }
        data: { $ref: '#/components/schemas/ValidationErrorDetail' }
...
    TestRunCreateApiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TestRunCreateRes' } }
    TestRunListApiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TestRunListRes' } }
...
    SuccessEnvelope:
      type: object
      required: [httpStatus, message, data]
      properties:
        httpStatus: { type: integer, minimum: 200, maximum: 299 }
        message: { type: string }
        data: { type: object }